plataforma/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
httpx==0.28.1
python-dotenv==1.1.0
openai==1.73.0
anthropic==0.50.0
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.12
chromadb==0.6.3
pypdf==5.4.0
tiktoken==0.9.0
